Better detection and handling of size
More specific selector for where the size is found on page, prevents
catching the 'up to 200 MB' allowed storage that advertises on every ZS
page.
Capture the size in the case when it's a round number, e.g. 20 MB was
ignored because it didn't look like a decimal (previous regex).
Use Number() instead of parseFloat(), as it's faster and automatically
gives us 0 when the string is empty, as it is for links to videos.